revert r23819: test more on postgresql after r23792
git-svn-id: https://svn.exedio.com/cope/persistence/trunk@23820 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/testsrc/com/exedio/cope/HierarchyTest.java
@@ -377,6 +377,9 @@ public class HierarchyTest extends AbstractRuntimeTest
 
 	public void testPrimaryKeyInfo()
 	{
+		if(postgresql) // causes a deadlock on postgresql
+			return;
+
 		MODEL.rollback();
 		// for flushing the info
 		MODEL.dropSchema();
